Adding views
We are adding the urlpatterns in urls.py from todo list so that all the urls are redirected to the urls.py in home directory.

We create an urls.py file, so that we can call the views by checking the pattern that urls.py from todolist has sent.

In views.py we add a new function called home, using which we return the desired behavior that we want to show on our website.